requirements.txt
gradio
web3
requests
python-docx
black
flake8
pre-commit
